pyproject.toml
[tool.poetry]
name = "python-decouple-aws"
version = "0.3.0"
description = "AWS Extensions for Python Decouple"
authors = ["Matt Magin <matt.magin@cmv.com.au>"]
license = "MIT"
repository = "https://github.com/AzMoo/python-decouple-aws"
readme = "README.rst"
packages = [
    { include = "decouple_aws" },
]

[tool.poetry.dependencies]
python = "^3.4"
python-decouple = "^3.1"
boto3 = "^1.7"

[tool.poetry.dev-dependencies]
prospector = "^1.1"
pre-commit = "^1.10"
pytest = "^3.2"
tox = "^3.4"
mock = "^2.0"
